@@ -14,8 +14,9 @@
     "html-webpack-plugin": "^5.5.0",
     "less": "^4.1.3",
     "less-loader": "^11.0.0",
+    "mini-css-extract-plugin": "^2.6.1",
     "style-loader": "^3.3.1",
-    "vue": "2.5.2",
+    "vue": "^3.2.36",
     "vue-loader": "^17.0.0",
     "vue-template-compiler": "^2.6.14",
     "webpack": "^5.73.0",
@@ -23,7 +24,10 @@
     "webpack-dev-server": "^4.9.2"
   },
   "scripts": {
-    "build": "webpack",
+    "build": "webpack --mode production",
     "serve": "webpack serve"
+  },
+  "dependencies": {
+    "install": "^0.13.0"
   }
 }
