@@ -1,5 +1,7 @@
 <template>
-  <div></div>
+  <div>
+    <h1>test</h1>
+  </div>
 </template>
 
 <script>
@@ -7,4 +9,7 @@ export default {}
 </script>
 
 <style lang="less" scoped>
+h1 {
+  color: hotpink;
+}
 </style>
